+++ frontend/src/components/HelpButton.js
@@ -1,6 +1,9 @@
 import React from "react";
 import HelpIcon from "@material-ui/icons/Help";
 import IconButton from "@material-ui/core/IconButton";
+import Box from "@material-ui/core/Box";
+import Typography from "@material-ui/core/Typography";
+import Modal from "@material-ui/core/Modal";
 
 /**
  * Material UI IconButton for representing a help/info button.
@@ -11,15 +14,50 @@ import IconButton from "@material-ui/core/IconButton";
  *   <HelpButton />
  * )
  */
+
+const style = {
+  position: "absolute",
+  top: "50%",
+  left: "50%",
+  transform: "translate(-50%, -50%)",
+  width: 400,
+  bgcolor: "background.paper",
+  border: "2px solid #000",
+  boxShadow: 24,
+  p: 4,
+};
+
 const HelpButton = () => {
+  const [open, setOpen] = React.useState(false);
+
+  const handleOpen = () => setOpen(true);
+  const handleClose = () => setOpen(false);
+
   return (
-    <IconButton
-      aria-label="help"
-      onClick={() => console.log("Link help popup here!")}
-      style={{ color: "white" }}
-    >
-      <HelpIcon />
-    </IconButton>
+    <>
+      <IconButton
+        aria-label="help"
+        onClick={handleOpen}
+        style={{ color: "white" }}
+      >
+        <HelpIcon />
+      </IconButton>
+      <Modal
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="modal-modal-title"
+        aria-describedby="modal-modal-description"
+      >
+        <Box sx={style}>
+          <Typography id="modal-modal-title" variant="h6" component="h2">
+            Help Section
+          </Typography>
+          <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+            Duis mollis, est non commodo luctus, nisi erat porttitor ligula.
+          </Typography>
+        </Box>
+      </Modal>
+    </>
   );
 };
